refactored buildAssembly()
